aop JDKProxy / CGLib example
