--- .editorconfig
@@ -1,18 +1,0 @@
-# top-most EditorConfig file
-root = true
-
-[*]
-indent_style = space
-indent_size = 2
-tab_width = 2
-max_line_length = 150
-charset = utf-8
-end_of_line = lf
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-# noinspection EditorConfigKeyCorrectness
-continuation_indent_size = 4
-
-[Makefile]
-indent_style = tab
