--- .gitattributes
@@ -1,2 +1,0 @@
-*.sh   text eol=lf
-*.bat  text eol=crlf
