--- .gitignore
@@ -1,5 +1,0 @@
-/logs/
-
-/target/
-/pom.xml.versionsBackup
-
