--- .travis.yml
@@ -1,11 +1,0 @@
-language: java
-jdk: openjdk14
-# install: true
-cache:
-  directories:
-    - /.m2
-script:
-  - mvn -N io.takari:maven:wrapper -Dmaven=3.6.1
-  - ./mvnw verify
-after_success:
-  - bash <(curl -s https://codecov.io/bash)
