--- jitpack.yml
@@ -1,2 +1,0 @@
-jdk:
-  - openjdk14
