--- lombok.config
@@ -1,11 +1,0 @@
-config.stopBubbling = true
-lombok.accessors.chain = true
-lombok.fieldDefaults.defaultPrivate = true
-lombok.fieldDefaults.defaultFinal = true
-lombok.addLombokGeneratedAnnotation = true
-lombok.extern.findbugs.addSuppressFBWarnings = true
-
-lombok.copyableAnnotations += org.springframework.beans.factory.annotation.Value
-#lombok.copyableAnnotations += com.fasterxml.jackson.annotation.JsonIgnore
-#lombok.copyableAnnotations += com.fasterxml.jackson.annotation.JsonProperty
-#lombok.copyableAnnotations += com.fasterxml.jackson.annotation.JsonValue
