--- ruleset.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ruleset xmlns="http://mojo.codehaus.org/versions-maven-plugin/rule/2.0.0"
-         comparisonMethod="maven"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="
-          http://mojo.codehaus.org/versions-maven-plugin/rule/2.0.0
-          http://www.mojohaus.org/versions-maven-plugin/xsd/rule-2.0.0.xsd">
-
-  <rules>
-    <rule groupId="org.codehaus.mojo" artifactId="sonar-maven-plugin" comparisonMethod="maven">
-      <ignoreVersions>
-        <ignoreVersion type="regex">3.*</ignoreVersion>
-      </ignoreVersions>
-    </rule>
-  </rules>
-
-  <ignoreVersions>
-    <ignoreVersion type="regex">.*alpha.*</ignoreVersion>
-    <ignoreVersion type="regex">.*beta.*</ignoreVersion>
-    <ignoreVersion type="regex">.*\.M\d+</ignoreVersion>
-  </ignoreVersions>
-
-</ruleset>
