+++ src/main/java/com/example/aop/service/Serv.java
@@ -1,0 +1,7 @@
+package com.example.aop.service;
+
+public interface Serv {
+
+  String math(int x);
+
+}
