+++ src/main/java/com/example/aop/service/Serv2.java
@@ -1,0 +1,14 @@
+package com.example.aop.service;
+
+import com.example.aop.aspect.MyAnno;
+import org.springframework.stereotype.Component;
+
+@Component
+public class Serv2 implements Serv {
+
+    @MyAnno
+    @Override
+    public String math(int x) {
+        return null;
+    }
+}
