+++ src/main/java/com/example/aop/service/ServImpl.java
@@ -1,0 +1,14 @@
+package com.example.aop.service;
+
+import com.example.aop.aspect.MyAnno;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ServImpl{
+
+    @MyAnno
+    public String math(int x) {
+        return Integer.toString(x);
+    }
+
+}
