--- src/main/java/ru/vlapin/experiments/demoproject/configs/Feign.java
@@ -1,13 +1,0 @@
-package ru.vlapin.experiments.demoproject.configs;
-
-import static org.springframework.hateoas.config.EnableHypermediaSupport.HypermediaType.HAL;
-
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.hateoas.config.EnableHypermediaSupport;
-
-@Configuration
-@EnableHypermediaSupport(type = HAL)
-@EnableFeignClients("ru.vlapin.experiments.demoproject.services")
-public class Feign {
-}
