--- src/main/java/ru/vlapin/experiments/demoproject/model/lombok.config
@@ -1,1 +1,0 @@
-lombok.fieldDefaults.defaultFinal = false
