+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+#define spring-boot AopAutoConfiguration proxy mode - default CGLIb
+spring.aop.proxy-target-class=false
