--- src/test/java/ru/vlapin/experiments/demoproject/DemoProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.vlapin.experiments.demoproject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoProjectApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
